:recycle: Move all cleanup subtasks into cleanup task definition
