+++ src/tasks/build.ts
@@ -1,8 +1,5 @@
-import * as del from 'del';
-import * as path from 'path';
 import * as sourcemaps from 'gulp-sourcemaps';
 import * as ts from 'gulp-typescript';
-import * as vinylPaths from 'vinyl-paths';
 
 import {initializeTypeScriptOptions} from './../setup/typescript-options';
 
@@ -15,24 +12,6 @@ export function generate(gulp, config, gulptraum): void {
     return gulp
       .src(allSourceFiles);
   }
-
-  const buildOutputFolderPath = path.resolve(config.paths.root, config.paths.output);
-
-  gulptraum.task('build-typescript-clean', {
-    help: 'Cleans the files compiled from your TypeScript source code'
-  }, () => {
-    // NOTE: Glob.js now throws an error by default, if a directly was not found.
-    // We must pass a config to "del", telling glob.js not to do that.
-    const deleteFiles: any = (patterns) => {
-      return del(patterns, {
-        nonull: false,
-      })
-    };
-
-    return gulp
-      .src(`${buildOutputFolderPath}`)
-      .pipe(vinylPaths(deleteFiles));
-  });
 
   config.compileToModules.forEach((moduleType) => {
