+++ src/tasks/doc.ts
@@ -1,6 +1,4 @@
 import * as path from 'path';
-import * as del from 'del';
-import * as vinylPaths from 'vinyl-paths';
 import * as typedoc from 'gulp-typedoc';
 import * as through2 from 'through2';
 
@@ -32,22 +30,6 @@ export function generate(gulp, config, gulptraum): void {
       return `lib.${libName}.d.ts`;
     })
   }
-
-  gulptraum.task('doc-typescript-clean', {
-    help: 'Cleans all documentation files generated by the TypeScript plugin'
-  }, () => {
-    // NOTE: Glob.js now throws an error by default, if a directly was not found.
-    // We must pass a config to "del", telling glob.js not to do that.
-    const deleteFiles: any = (patterns) => {
-      return del(patterns, {
-        nonull: false,
-      })
-    };
-
-    return gulp
-      .src(`${docsOutputFolderPath}`)
-      .pipe(vinylPaths(deleteFiles));
-  });
 
   gulptraum.task('doc-typescript-generate', {
     help: 'Generates the documentation from your TypeScript source code using TypeDoc'
